Restructure More Colors add-on, refactor operators
Move add-on code into More_Colors package, and add registration modules. Rename operator classes to MC_OT_* and rename/refactor many operator implementations (add_random_color, add_color_by_position, simple_fill, reset/display/open docs) to improve readability and PEP8 styling. Add new PropertyGroup modules for tools and settings, hook them into Scene via PointerProperty, and add UI module scaffolding. Improve correctness and performance: avoid division-by-zero, call obj.data.update() after edits, reuse material lookups, build vert->loop and edge->face maps to avoid O(V×F)/O(E×F) patterns, apply global color masks, and add bl_options={'REGISTER','UNDO'} where appropriate. Bump blender_manifest.toml version to 1.2.0 and add utilities (logging) and other small fixes.
