+++ .gitignore
@@ -2,3 +2,4 @@ __pycache__
 .vscode
 *.blend
 *.blend1
+*.zip
