+++ More_Colors/blender_manifest.toml
@@ -1,0 +1,17 @@
+schema_version = "1.0.0"
+
+id = "more_colors"
+version = "1.2.0"
+name = "More Colors!"
+tagline = "A set of tools to make vertex painting easier"
+maintainer = "Kai Fardreamer <tojynick@protonmail.com>"
+type = "add-on"
+
+website = "https://github.com/tojynick/More-Colors"
+tags = ["3D View", "Mesh", "Paint"]
+
+blender_version_min = "4.2.0"
+
+license = [
+  "SPDX:GPL-3.0-or-later"
+]
