+++ More_Colors/ui/base_panel_info.py
@@ -1,0 +1,12 @@
+# Copyright (C) 2024 Kai Fardreamer <tojynick@protonmail.com>
+# Copyright (C) 2026 Clonephaze
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+
+class BasePanelInfo:
+    """Contains basic panel properties like space and region types, as well as a category."""
+
+    bl_space_type = "VIEW_3D"
+    bl_region_type = "UI"
+    bl_category = "More Colors!"
