--- ui/base_panel_info.py
@@ -1,5 +1,0 @@
-class BasePanelInfo:
-    """ Containts basic panel properties like space and region types, as well as a category."""
-    bl_space_type = "VIEW_3D"
-    bl_region_type = "UI"
-    bl_category = "More Colors!"
